Convert RecordTimeStepperData to a mutator
