+++ src/Evolution/Systems/Cce/Components/KleinGordonCharacteristicEvolution.hpp
@@ -14,11 +14,13 @@
 #include "Parallel/GlobalCache.hpp"
 #include "Parallel/Local.hpp"
 #include "Parallel/Phase.hpp"
+#include "ParallelAlgorithms/Actions/MutateApply.hpp"
 #include "Time/Actions/CleanHistory.hpp"
 #include "Time/Actions/SelfStartActions.hpp"
 #include "Time/Actions/TakeLtsStep.hpp"
 #include "Time/Actions/UpdateU.hpp"
 #include "Time/ChangeTimeStepperOrder.hpp"
+#include "Time/RecordTimeStepperData.hpp"
 #include "Utilities/TMPL.hpp"
 
 namespace Cce {
@@ -117,7 +119,7 @@ struct KleinGordonCharacteristicEvolution
       tmpl::transform<typename metavariables::cce_scri_tags,
                       tmpl::bind<::Actions::MutateApply,
                                  tmpl::bind<CalculateScriPlusValue, tmpl::_1>>>,
-      ::Actions::RecordTimeStepperData<cce_system>,
+      ::Actions::MutateApply<RecordTimeStepperData<cce_system>>,
       ::Actions::UpdateU<cce_system, Metavariables::local_time_stepping>>;
 
   using extract_action_list = tmpl::list<
