+++ src/Time/Actions/CMakeLists.txt
@@ -7,7 +7,6 @@ spectre_target_headers(
   HEADERS
   AdvanceTime.hpp
   CleanHistory.hpp
-  RecordTimeStepperData.hpp
   SelfStartActions.hpp
   TakeLtsStep.hpp
   UpdateU.hpp
