+++ src/Time/TakeStep.hpp
@@ -4,9 +4,9 @@
 #pragma once
 
 #include "DataStructures/DataBox/DataBox.hpp"
-#include "Time/Actions/RecordTimeStepperData.hpp"
 #include "Time/Actions/UpdateU.hpp"
 #include "Time/ChangeStepSize.hpp"
+#include "Time/RecordTimeStepperData.hpp"
 #include "Utilities/Gsl.hpp"
 
 /// Bundled method for recording the current system state in the history, and
@@ -17,6 +17,6 @@ void take_step(const gsl::not_null<db::DataBox<DbTags>*> box) {
   if constexpr (LocalTimeStepping) {
     db::mutate_apply<ChangeStepSize<StepChoosersToUse>>(box);
   }
-  record_time_stepper_data<System>(box);
+  db::mutate_apply<RecordTimeStepperData<System>>(box);
   update_u<System>(box);
 }
