+++ tests/Unit/Time/Actions/CMakeLists.txt
@@ -5,7 +5,6 @@ set(LIBRARY_SOURCES
   ${LIBRARY_SOURCES}
   Actions/Test_AdvanceTime.cpp
   Actions/Test_CleanHistory.cpp
-  Actions/Test_RecordTimeStepperData.cpp
   Actions/Test_SelfStartActions.cpp
   Actions/Test_TakeLtsStep.cpp
   Actions/Test_UpdateU.cpp
