+++ tests/Unit/Time/CMakeLists.txt
@@ -13,6 +13,7 @@ set(LIBRARY_SOURCES
   Test_EvolutionOrdering.cpp
   Test_History.cpp
   Test_LargestStepperError.cpp
+  Test_RecordTimeStepperData.cpp
   Test_SelfStart.cpp
   Test_Slab.cpp
   Test_SlabRoundingError.cpp
